+++ extensions/gif-search/CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add custom Empty State
 - Add preference for max number of search results (defaults to 10)
 - Add Favorites view so you can see and search all your favorite gifs across all services
+- Remove separate service commands in favor of the one unified "Search for GIFs" command
 
 ## [Remove static GIF preview warning 🎉] - 2022-04-06
 - Removes the warning about GIFs no longer animating, Raycast v1.32.0 supports GIFs in Details